combine separated files back into powerlinx.go
